refactor: split properties/bookings/services into sub-modules
- api/properties/* — crud, reviews, catalog, availability, ical
- api/bookings/*  — queries, mutations
- api/services/*  — crud, catalog, edits
- Original files become thin re-export wrappers (backward compat)
- Zero changes to external imports or public API surface
