+++ .mcp.json
@@ -1,0 +1,15 @@
+{
+  "mcpServers": {
+    "supabase": {
+      "type": "http",
+      "url": "https://mcp.supabase.com/mcp?project_ref=mdmizeyiyebvhkujjyjg"
+    },
+    "notion": {
+      "command": "npx",
+      "args": ["-y", "@notionhq/notion-mcp-server"],
+      "env": {
+        "OPENAPI_MCP_HEADERS": "{\"Authorization\": \"Bearer ntn_m5654676453WSycVux6hgVRrgifxy8r1tXslIbLouA63Do\", \"Notion-Version\": \"2022-06-28\"}"
+      }
+    }
+  }
+}
