+++ api-services/api/properties/index.ts
@@ -1,0 +1,49 @@
+import {
+    getPropertiesByIds, createProperty, getProperties, getAvailableProperties,
+    getProperty, getPropertiesByHost, getAdminProperties,
+    updateProperty, updatePropertyStatus, deleteProperty
+} from './crud';
+import { getReviews, getReviewCount, addReview, deleteReview } from './reviews';
+import { getPropertyTypes, getPropertyLocations, getPropertiesByLocation } from './catalog';
+import {
+    getPropertyAvailability, updatePropertyAvailability,
+    syncPropertyCalendar, getUnavailableDates
+} from './availability';
+import { getICalFeeds, addICalFeed, removeICalFeed } from './ical';
+
+export const propertiesService = {
+    // CRUD
+    getPropertiesByIds,
+    createProperty,
+    getProperties,
+    getAvailableProperties,
+    getProperty,
+    getPropertiesByHost,
+    getAdminProperties,
+    updateProperty,
+    updatePropertyStatus,
+    approveProperty: (id: string) => updatePropertyStatus(id, 'approved'),
+    deleteProperty,
+
+    // Reviews
+    getReviews,
+    getReviewCount,
+    addReview,
+    deleteReview,
+
+    // Catalog
+    getPropertyTypes,
+    getPropertyLocations,
+    getPropertiesByLocation,
+
+    // Availability
+    getPropertyAvailability,
+    updatePropertyAvailability,
+    syncPropertyCalendar,
+    getUnavailableDates,
+
+    // iCal
+    getICalFeeds,
+    addICalFeed,
+    removeICalFeed,
+};
